requirements.txt
torch>=2.0.0
numpy>=1.21.0
matplotlib>=3.5.0
tqdm>=4.60.0
networkx>=2.8.0
scikit-learn>=1.0.0
jupyter>=1.0.0
h5py>=3.0.0
umap-learn>=0.5.0
plotly>=5.0.0 